+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[13.7.2]
+- FABRIC: Made dependency on Forge Config API Port optional
+    - If the mod is not installed, config loading will not be supported
+
 [13.7.1]
 - Fix crash when using a remote upgrade